+++ impl_core/src/main/java/io/opencensus/implcore/stats/MutableAggregation.java
@@ -33,9 +33,13 @@ import io.opencensus.stats.BucketBoundaries;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 /** Mutable version of {@link Aggregation} that supports adding values. */
 abstract class MutableAggregation {
+
+  private static final Logger logger = Logger.getLogger(MutableAggregation.class.getName());
 
   private MutableAggregation() {}
 
@@ -292,6 +296,11 @@ abstract class MutableAggregation {
 
     @Override
     void add(double value, Map<String, String> attachments, Timestamp timestamp) {
+      if (value < 0) {
+        logger.log(Level.WARNING, "Dropping negative values for histogram count and sum.");
+        return;
+      }
+
       sum += value;
       count++;