+++ impl_core/src/test/java/io/opencensus/implcore/stats/MutableAggregationTest.java
@@ -139,7 +139,8 @@ public class MutableAggregationTest {
         TOLERANCE);
     assertAggregationDataEquals(
         aggregations.get(4).toAggregationData(),
-        AggregationData.DistributionData.create(4.0, 5, -5.0, 20.0, 372, Arrays.asList(2L, 2L, 1L)),
+        AggregationData.DistributionData.create(
+            8.666666, 3, 1.0, 20.0, 200.666666, Arrays.asList(0L, 2L, 1L)),
         TOLERANCE);
     assertAggregationDataEquals(
         aggregations.get(5).toAggregationData(),
@@ -180,7 +181,7 @@ public class MutableAggregationTest {
     // bucket, only the last one will be kept.
     List<Exemplar> expected =
         Arrays.<Exemplar>asList(
-            Exemplar.create(values.get(2), timestamps.get(2), attachmentsList.get(2)),
+            null,
             Exemplar.create(values.get(4), timestamps.get(4), attachmentsList.get(4)),
             Exemplar.create(values.get(3), timestamps.get(3), attachmentsList.get(3)));
     assertThat(mutableDistribution.getExemplars())
@@ -256,13 +257,14 @@ public class MutableAggregationTest {
     MutableDistribution combined = MutableDistribution.create(BUCKET_BOUNDARIES);
     combined.combine(distribution1, 1.0); // distribution1 will be combined
     combined.combine(distribution2, 0.6); // distribution2 will be ignored
-    verifyMutableDistribution(combined, 0, 2, -5, 5, 50.0, new long[] {1, 1, 0}, TOLERANCE);
+    verifyMutableDistribution(combined, 5, 1, 5, 5, 0.0, new long[] {0, 1, 0}, TOLERANCE);
 
     combined.combine(distribution2, 1.0); // distribution2 will be combined
-    verifyMutableDistribution(combined, 7.5, 4, -5, 20, 325.0, new long[] {1, 1, 2}, TOLERANCE);
+    verifyMutableDistribution(
+        combined, 11.666666, 3, 5, 20, 116.66666666, new long[] {0, 1, 2}, TOLERANCE);
 
     combined.combine(distribution3, 1.0); // distribution3 will be combined
-    verifyMutableDistribution(combined, 0, 8, -20, 20, 1500.0, new long[] {4, 1, 3}, TOLERANCE);
+    verifyMutableDistribution(combined, 12.5, 4, 5, 20, 125.0, new long[] {0, 1, 3}, TOLERANCE);
   }
 
   @Test